rsx: Separate register context from RSX thread
