+++ rpcs3/Emu/RSX/NV47/context_accessors.undef.h
@@ -1,0 +1,3 @@
+#undef RSX
+#undef REGS
+#undef RSX_CAPTURE_EVENT
