+++ rpcs3/Emu/RSX/NV47/nv3089.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "context.h"
+
+namespace rsx
+{
+	namespace nv3089
+	{
+		void image_in(context* ctx, u32 reg, u32 arg);
+	}
+}
