+++ rpcs3/Emu/RSX/NV47/nv47.h
@@ -1,0 +1,7 @@
+// 3D Engine definitions
+#pragma once
+
+#include "nv3089.h"
+#include "nv308a.h"
+#include "nv406e.h"
+#include "nv4097.h"
