+++ rpcs3/emucore.vcxproj
@@ -95,6 +95,11 @@
     <ClCompile Include="Emu\perf_monitor.cpp" />
     <ClCompile Include="Emu\RSX\Common\texture_cache.cpp" />
     <ClCompile Include="Emu\RSX\Core\RSXContext.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\common.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\nv3089.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\nv308a.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\nv406e.cpp" />
+    <ClCompile Include="Emu\RSX\NV47\nv4097.cpp" />
     <ClCompile Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu.cpp" />
     <ClCompile Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_components.cpp" />
     <ClCompile Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_message_box.cpp" />
@@ -584,6 +589,15 @@
     <ClInclude Include="Emu\RSX\Core\RSXDisplay.h" />
     <ClInclude Include="Emu\RSX\Core\RSXReservationLock.hpp" />
     <ClInclude Include="Emu\RSX\Core\RSXVertexTypes.h" />
+    <ClInclude Include="Emu\RSX\NV47\context.h" />
+    <ClInclude Include="Emu\RSX\NV47\context_accessors.define.h" />
+    <ClInclude Include="Emu\RSX\NV47\context_accessors.undef.h" />
+    <ClInclude Include="Emu\RSX\NV47\nv3089.h" />
+    <ClInclude Include="Emu\RSX\NV47\nv308a.h" />
+    <ClInclude Include="Emu\RSX\NV47\nv406e.h" />
+    <ClInclude Include="Emu\RSX\NV47\nv4097.h" />
+    <ClInclude Include="Emu\RSX\NV47\nv47.h" />
+    <ClInclude Include="Emu\RSX\NV47\common.h" />
     <ClInclude Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu.h" />
     <ClInclude Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_components.h" />
     <ClInclude Include="Emu\RSX\Overlays\HomeMenu\overlay_home_menu_message_box.h" />
